HBASE-9329 SnapshotManager should check for directory existance before throwing a warning (Jean-Marc Spaggiari)
git-svn-id: https://svn.apache.org/repos/asf/hbase/branches/0.94@1517609 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/hadoop/hbase/master/snapshot/SnapshotManager.java
@@ -263,8 +263,10 @@ public class SnapshotManager implements Stoppable {
   void resetTempDir() throws IOException {
     // cleanup any existing snapshots.
     Path tmpdir = SnapshotDescriptionUtils.getWorkingSnapshotDir(rootDir);
-    if (!master.getMasterFileSystem().getFileSystem().delete(tmpdir, true)) {
-      LOG.warn("Couldn't delete working snapshot directory: " + tmpdir);
+    if (master.getMasterFileSystem().getFileSystem().exists(tmpdir)) {
+      if (!master.getMasterFileSystem().getFileSystem().delete(tmpdir, true)) {
+        LOG.warn("Couldn't delete working snapshot directory: " + tmpdir);
+      }
     }
   }
